refactor: add shared colors/helpers to zp_utils; convert build_zpj.py
